eliminate editor import with browser support check
